+++ src/client/js/index.js
@@ -11,3 +11,5 @@ import './current.js';
 import './edit_panel.js';
 import './startup.js';
 import './view_add.js';
+//
+import '../stylesheets/roles.less';