+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Define Roles.usedScopes() new function, bumping minor candidate version number
     - Roles.configure() becomes a reactive data source
     - Introduce trace module
+    - Introduce 'pr-edit-state' new event
 
 ### 1.4.0